requirements.txt
coverage==7.4.4
iniconfig==2.0.0
packaging==24.0
pluggy==1.4.0
pytest==8.1.1
pytest-cov==4.1.0
python-dotenv==1.0.1
